+++ deployment/updates/scripts/add_permissions/2015_07_29_allow_user_session_uiconf_listTemplates.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package deployment
+ * @subpackage jupiter.roles_and_permissions
+ */
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$config = realpath(dirname(__FILE__)) . '/../../../permissions/service.uiconf.ini';
+passthru("php $script $config");