+++ release-notes.md
@@ -1,5 +1,22 @@
 # Jupiter-10.17.0 #
 
+## Allow users to use uiconf-->listTemplates ##
+
+- Issue Type: Permission to existing API 
+- Issue ID: PLAT-3541
+
+#### Configuration ####
+ 
+ 	- In the file "deployment/permissions/service.uiconf.ini" in the line "permissionItem7.permissions" add `,BASE_USER_SESSION_PERMISSION` at the end
+
+#### Deployment Scripts ####
+
+ 	- php deployment/updates/scripts/add_permissions/2015_07_29_allow_user_session_uiconf_listTemplates.php
+
+#### Known Issues & Limitations ####
+
+None.
+
 ## new Transcript asset ##
 
 -- Issue Type: New Feature
@@ -20,8 +37,8 @@ None.
 
 ## Allow PLAYBACK_BASE_ROLE to user user_entry and quiz ##
 
--- Issue Type: Permission to existing API 
--- Issue ID:
+- Issue Type: Permission to existing API 
+- Issue ID:
 
 #### Configuration ####